Get cat and post cat

+++ Data/EFCore/EFCoreRepository.cs
@@ -18,7 +18,7 @@ namespace catsapi.Data.EFCore
             return entity;
         }
 
-        public async Task<TEntity> Delete(int id)
+        public async Task<TEntity> Delete(long id)
         {
             var entity = await context.Set<TEntity>().FindAsync(id);
             if (entity == null)
@@ -32,7 +32,7 @@ namespace catsapi.Data.EFCore
             return entity;
         }
 
-        public async Task<TEntity> Get(int id)
+        public async Task<TEntity> Get(long id)
         {
             return await context.Set<TEntity>().FindAsync(id);
         }
