+++ Data/IRepository.cs
@@ -6,9 +6,9 @@ namespace catsapi.Data
     public interface IRepository<T> where T : class, IEntity
     {
         Task<List<T>> GetAll();
-        Task<T> Get(int id);
+        Task<T> Get(long id);
         Task<T> Add(T entity);
         Task<T> Update(T entity);
-        Task<T> Delete(int id);
+        Task<T> Delete(long id);
     }
 }
